Reverted GetTickCount64 patch

+++ main/Helper.h
@@ -64,5 +64,5 @@ int getclock(struct timeval *tv);
 int timeval_subtract (struct timeval *result, struct timeval *x, struct timeval *y);
 
 bool IsArgumentSecure(const std::string &arg);
-uint64_t SystemUptime();
+uint32_t SystemUptime();
